Commit Música
En este commit he metido la música, va por escenas y se corta de un menú a otro. No me desagrada del todo pero puede resultar molesto a la hora de jugar. Por otro lado sigo trabajando con las colisiones, el box collider del castle_prefab no me responde muy bien.
