+++ Assets/Materials/sea.mat
@@ -73,6 +73,6 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0, g: 0.044748515, b: 0.1981132, a: 1}
+    - _Color: {r: 0.06888572, g: 0.12710865, b: 0.33962262, a: 1}
     - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
   m_BuildTextureStacks: []
