+++ Assets/Materials/sky.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c1bf7192b825d04ebeced2c9be4bf5d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
