+++ Assets/Prefabs/castle.prefab
@@ -125,8 +125,8 @@ BoxCollider:
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 11, y: 40, z: 11}
-  m_Center: {x: 0, y: 20.2, z: 0}
+  m_Size: {x: 11.6, y: 42.06, z: 12.369333}
+  m_Center: {x: -0.31171113, y: 21.288958, z: -0.025413515}
 --- !u!114 &4689997008814558574
 MonoBehaviour:
   m_ObjectHideFlags: 0
