+++ Assets/Scripts/CollisionManager.cs
@@ -4,44 +4,47 @@ using UnityEngine;
 
 public class CollisionManager : MonoBehaviour
 {
-    
-    
+    public static int lives;
     private PlayerMovement playerMovement;
 
     // Start is called before the first frame update
     void Start()
     {
-       //alive = true;
-      // PlayerMovement.lives = 3;
+
     }
 
     // Update is called once per frame
     void Update()
     {
 
-    
-    }
-    /*
-    void Die()
-    {
-        if (lives == 0)
-        {
-            Die();
-        }
-    }
-    
-    private void OnTriggerEnter(Collider other)
-    {
-        {
-            print(PlayerMovement.lives);
-            print(other.tag);
-            if (other.tag == "Prefab")
-            {
-                lives--;
-            }
-            print(lives);
-        }
 
     }
-    */
+    public class ConfigVars : MonoBehaviour
+    {
+        void Die()
+        {
+            if (lives == 0)
+            {
+                Die();
+            }
+        }
+
+        void OnTriggerEnter(Collider other)
+        {
+            {
+                print(PlayerMovement.lives);
+                print(other.tag);
+                if (other.tag == "Prefab")
+                {
+                    lives--;
+                }
+                print(lives);
+               
+            }
+
+        }
+    }
+ 
+     
+    
 }
