+++ Assets/Scripts/LightManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e81587f90c3661242899f3989421995e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
