+++ Assets/Scripts/PlayerMovement.cs
@@ -4,21 +4,25 @@ using UnityEngine;
 
 public class PlayerMovement : MonoBehaviour
 {
-    //Vamos a realizar el movimiento por físicas por lo tanto le damos un rigidbody al Player
+    //Vidas
+    public static int lives;
+
+    //Rigidbody al Player.Movimiento por físicas.
     [SerializeField] float strength;
     [SerializeField] Rigidbody dragonRigidBody;
 
-    //límites de movimiento
-    //public float horizontalLimit = 100f;
-    public float verticalLimit = 10f;
+    //límites de movimiento vertical
+    public float verticalLimit = 30f;
 
     //variable alive
-    // dato que nos viene bien saber para la gestión de prefabs, dado que dependerán de ello para existir
-    public bool alive;
-    public PlayerMovement playerMovement;
-    [SerializeField]public int lives;
+    // dato que nos viene bien saber para la gestión de prefabs, 
+    // ya que dependerán de ello para existir.
 
-    public float dragonSpeed = 0.5f;
+    public static bool alive;
+    public PlayerMovement playerMovement;
+
+    //Velocidad
+    public float dragonSpeed = 0.2f;
     
 
 
@@ -29,18 +33,18 @@ public class PlayerMovement : MonoBehaviour
         //Le damos el valor true a alive, comienza el juego
         alive = true;
         lives = 3;
-       // for (float dragonSpeed = 0; dragonSpeed < 10; dragonSpeed += 0.0000001f);
+       
     }
 
     // Update is called once per frame
     void Update()
     {
         
-        
+        //Solo si estoy vivo, déjame moverme.
         if (alive == true)
         {
             playerPhysicsMovement();
-            dragonSpeed += 0.000001f;
+            dragonSpeed += 0.0001f;
 
         }
     }
@@ -68,43 +72,22 @@ public class PlayerMovement : MonoBehaviour
             if (Input.GetKeyDown(KeyCode.Space))
 
         {
-            GetComponent<Rigidbody>().AddForce(Vector3.up * 3000);
+            GetComponent<Rigidbody>().AddForce(Vector3.up * 4000);
 
         }
      
 
     }
-    
 
-
-
-    /*
     //Método para que muera la nave
     void Die()
     {
-        if(ConfigVars.numLives == 0)
+        if(lives == 0)
         {
-            Die();
+            alive = false;
+            Destroy(this.gameObject,3f);
         }
     }
-    */
-    /*
-       if ((transform.position.x >= -horizontalLimit) && (transform.position.x <= horizontalLimit))
-       {
-           //movimiento lateral según fuerzas
-           dragonRigidBody.AddForce(Vector3.right * desplH * strength);
-           dragonRigidBody = GetComponent<Rigidbody>();
-       }
-       */
-    //de alguna manera esta condición debe resolver que se corrija el movimiento si se detecta un valor válido
-    //quizás podemos hacer que al pasar el límite se nos empuje hacia el origen
-    //no se produce el impulso, creo que es porque no detecta nunca que llegue al límite
-    /*
-    if ((transform.position.x <= -horizontalLimit) && (transform.position.x >= horizontalLimit))
-    {
-        dragonRigidBody.AddForce(Vector3.right * desplH * strength * 4000);
-        dragonRigidBody = GetComponent<Rigidbody>();
+    
 
-    }
-    */
 }
