+++ UserSettings/EditorUserSettings.asset
@@ -6,31 +6,31 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 22424703114646680e0b0227036c65312e2f370413040f1b08162916d1d40776f7e93ffdfe
-      flags: 0
-    RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c65312e2f370413040f1b08162f16d6d20d07cfc218dca92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-2:
+    RecentlyUsedScenePath-1:
       value: 22424703114646680e0b0227036c65312e2f370413040f1b08163a12cfc50b14c7d113c5a92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-3:
+    RecentlyUsedScenePath-2:
       value: 22424703114646680e0b0227036c65312e2f370413040f1b08163203d6c91b16d1d81bccc90f7129370bfb25
       flags: 0
-    RecentlyUsedScenePath-4:
+    RecentlyUsedScenePath-3:
       value: 22424703114646680e0b0227036c65312e2f370413040f1b08162e07c3d20007cfc218dca92f31352d1b
       flags: 0
-    RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c0f4047282b1e0d1a120f000c3306acf53a31f6fe
-      flags: 0
-    RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c0f404428371a1801091e1e163016ccd57a2decee22f0
-      flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c0f4045283908031d120f000c3306acf53a31f6fe
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c0f4046283f0b010d191c081f381facf53a31f6fe
+      flags: 0
+    RecentlyUsedScenePath-6:
+      value: 22424703114646680e0b0227036c0f4047282b1e0d1a120f000c3306acf53a31f6fe
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 22424703114646680e0b0227036c0f404428371a1801091e1e163016ccd57a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c0f4045283908031d120f000c3306ddae2136ebf32f
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c0f4046283f0b010d191c081f381fddc73535e7f73ae8fe742a323016f6
